project_l7q7p/src/main/persistence/JsonWriter.java
package persistence;

import model.MyEvent;
import org.json.JSONObject;

import java.io.FileNotFoundException;
import java.io.PrintWriter;

// The code in the class below is from the JsonWriter class,
// in the JsonSerializationDemo project given as an example.
// Represents a writer that writes JSON representation of the event to file
public class JsonWriter {
    private static final int TAB = 4;
    private PrintWriter writer;
    private final String destination;


    // EFFECTS: constructs writer to write to destination file
    public JsonWriter(String destination) {
        this.destination = destination;
    }

    // MODIFIES: this
    // EFFECTS: opens writer; throws FileNotFoundException if destination file cannot
    // be opened for writing
    public void open() throws FileNotFoundException {
        writer = new PrintWriter(destination);
    }

    // MODIFIES: this
    // EFFECTS: writes JSON representation of event to file
    public void write(MyEvent event) {
        JSONObject json = event.toJson();
        saveToFile(json.toString(TAB));
    }

    // MODIFIES: this
    // EFFECTS: closes writer
    public void close() {
        writer.close();
    }

    // MODIFIES: this
    // EFFECTS: writes string to file
    private void saveToFile(String json) {
        writer.print(json);
    }

}
